+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Core
 
 * Feature: Activated Apps lifecycle, enabling and disabling apps on the fly depending on it's requirements.
 * Feature: Add player toolbox/toolbar. You can disable this with a setting in //settings in-game.
+* Feature: Add CP Comparison to find the best checkpoints by using the best checkpoint times of all local records (/cpcomparison).
 
 * Improvement: Dropping Python 3.5 support!
 * Improvement: Add //helpall and /helpall for a detailed list of commands!